refactor: adjust hook exports
