+++ packages/hydrogen-sanity/src/image.ts
@@ -1,6 +1,6 @@
-import createImageUrlBuilder from '@sanity/image-url'
+import {default as createImageUrlBuilder} from '@sanity/image-url'
 import type {ImageUrlBuilder} from '@sanity/image-url/lib/types/builder'
-import type {SanityImageSource} from '@sanity/image-url/lib/types/types'
+import type {SanityImageSource, SanityModernClientLike} from '@sanity/image-url/lib/types/types'
 import {useMemo} from 'react'
 
 import {useSanityProviderValue} from './provider'
@@ -8,7 +8,9 @@ import {useSanityProviderValue} from './provider'
 export function useImageUrlBuilder(): ImageUrlBuilder {
   const {projectId, dataset, apiHost} = useSanityProviderValue()
   return useMemo(() => {
-    return createImageUrlBuilder({config: () => ({projectId, dataset, apiHost})})
+    return createImageUrlBuilder({
+      config: () => ({projectId, dataset, apiHost}),
+    } as SanityModernClientLike)
   }, [apiHost, dataset, projectId])
 }
