--- packages/hydrogen-sanity/src/preview/index.ts
@@ -1,1 +1,0 @@
-export {usePreviewMode} from './hooks'
